+++ src/demo_substrate_events/types/assethub/substrate_events/assets_transferred/v601.py
@@ -4,9 +4,6 @@ from __future__ import annotations
 
 from typing import TypedDict
 
-"""
-Some assets were transferred. [asset_id, from, to, amount]
-"""
 V601 = TypedDict(
     'V601',
     {